Transition logrotate::base to logrotate
This is necessary if you want to ensure logrotate is present but are not
adding any additional logrotate::rule{}'s. This goes from having a
private class to a public class, but keeps the original goal of not
having to include a separate class when using the logrotate::rule
defined type.